% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_clustering_labels}
\alias{get_clustering_labels}
\title{Get clustering labels (helper function)}
\usage{
get_clustering_labels(data, clustering_algo, algo_params, labels_name)
}
\arguments{
\item{data}{The dataset to cluster.}

\item{clustering_algo}{A function that returns cluster labels (vector).}

\item{algo_params}{A list of parameters passed to `clustering_algo`.}
}
\value{
Integer or numeric vector of cluster labels.
}
\description{
Get clustering labels (helper function)
}
