% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_optimization_of_parameter_boxplot}
\alias{plot_optimization_of_parameter_boxplot}
\title{Plot Optimization of Parameter (Boxplot)}
\usage{
plot_optimization_of_parameter_boxplot(
  optimize_parameters_results_ecc,
  name_parameter = "optimized Parameter",
  title = ""
)
}
\arguments{
\item{optimize_parameters_results_ecc}{A named list, where each name is a parameter value
and the corresponding element is a numeric vector of ECC values.}

\item{name_parameter}{A character string for the x-axis label (default="optimized Parameter").}

\item{title}{A character string for the plot title (default="").}
}
\value{
A ggplot object representing the boxplot.
}
\description{
Plot the optimization of a single parameter. Call this function after running parameter_optimizer.
}
