% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{parameter_searcher}
\alias{parameter_searcher}
\title{Full Grid Search for Stochastic Clustering Parameters}
\usage{
parameter_searcher(
  data,
  clustering_algo,
  labels_name,
  param_grid,
  n_runs = 30,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{A data matrix or data frame to cluster.}

\item{clustering_algo}{A function that accepts a seed parameter (named as in
\code{parameter_name_seed}) and returns cluster labels for each row of `data`.}

\item{labels_name}{A string with the name of the label containing the clusters returned by the algorithm.}

\item{param_grid}{A named list of form
\code{list(paramA = c(values), paramB = c(values), ...)}, specifying
all parameter values to be combined.}

\item{n_runs}{Number of repeated runs (defaults to 30).}

\item{verbose}{Logical; if \code{TRUE}, print progress messages.}

\item{...}{Additional parameters passed to \code{clustering_algo}.}
}
\value{
A list with two components:
\itemize{
  \item \code{results_df}: A data frame. Each row corresponds to one parameter
  combination and contains:
  \itemize{
    \item \code{params}: A (list-)column with a named list of parameter values
    \item \code{ecc}: The numeric vector of ECC values for that combination
    \item \code{median_ecc}: The median ECC
  }
  \item \code{stochastic_clustering_results}: a named list of the full
    outputs from \code{stochastic_clustering_runner} for each parameter
    combination, keyed by a string like \code{"paramA_valA_paramB_valB"}.
}
}
\description{
Takes a grid of parameter combinations and runs repeated stochastic clustering
for each combination, computing element consistency (ECC) in each case.
}
