% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ga_fitness_evolution}
\alias{plot_ga_fitness_evolution}
\title{Plot the best fitness evolution over generations.}
\usage{
plot_ga_fitness_evolution(ga_fs_history)
}
\arguments{
\item{ga_fs_history}{A data frame containing the history of the feature selection GA,
with columns "generation" and "best_fitness".}
}
\value{
A ggplot object representing the best fitness evolution plot.
}
\description{
Plots the best fitness evolution across generations from a genetic algorithm.
}
